+++ README.md
@@ -1,1 +1,10 @@
-# hagall-common
+# Hagall Common Libraries
+
+This repository contains packages that are commonly used by Hagal & Other backend services in Aukilabs.
+
+## Packages
+
+| Package            | Description                                                           |
+| ------------------ | --------------------------------------------------------------------- |
+| [cli](cli)         | Go package for building cli tool that can load program configuration from multiple sources. |
+| [crypt](crypt)     | Package handling cryptography related functionality. |
