+++ crypt/handler.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 	"net/http"
 
-	"github.com/aukilabs/hagall-common/errors"
-	"github.com/aukilabs/hagall-common/logs"
+	"github.com/aukilabs/go-tooling/pkg/errors"
+	"github.com/aukilabs/go-tooling/pkg/logs"
 )
 
 const (
