--- events/event.go
@@ -1,4 +1,0 @@
-package events
-
-// Event represents an event to send.
-type Event any
