+++ scenario/scenario.go
@@ -3,7 +3,7 @@ package scenario
 import (
 	"context"
 
-	"github.com/aukilabs/hagall-common/errors"
+	"github.com/aukilabs/go-tooling/pkg/errors"
 	"github.com/aukilabs/hagall-common/messages/hagallpb"
 	hwebsocket "github.com/aukilabs/hagall-common/websocket"
 	"golang.org/x/net/websocket"
