+++ smoketest/smoketest.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/aukilabs/hagall-common/errors"
+	"github.com/aukilabs/go-tooling/pkg/errors"
 	"github.com/aukilabs/hagall-common/messages/hagallpb"
 	"github.com/aukilabs/hagall-common/scenario"
 	hwebsocket "github.com/aukilabs/hagall-common/websocket"
